crud

+++ products/index.php
@@ -2,7 +2,7 @@
 require_once '../templates/header.php';
 ?>
 
-<a href="/fullstack-php-beginner/products/create.php">
+<a href="/php-beginner/products/create.php">
     <button type="button">+Product</button>
 </a>
 
@@ -12,11 +12,11 @@ require_once '../templates/header.php';
         <th>Name</th>
         <th>Price</th>
         <th>Weight</th>
-        <th>Description</th>
-        <th>Thumbnail</th>
         <th>Discount</th>
         <th>stock</th>
         <th>Category</th>
+        <th>Description</th>
+        <th>Thumbnail</th>
         <th>🎮</th>
     </tr>
 
@@ -29,15 +29,15 @@ require_once '../templates/header.php';
                 <td>$product->name</td>
                 <td>$product->price</td>
                 <td>$product->weight</td>
-                <td>$product->description</td>
-                <td>$product->thumbnail</td>
                 <td>$product->discount</td>
                 <td>$product->stock</td>
                 <td>". ($product->category_fk=='' ? '-':$product->category_fk) ."</td>
+                <td>$product->description</td>
+                <td>$product->thumbnail</td>
                 <td>
                     <a href='./detail.php?id=$product->id'>Detail</a>   
                     <a href='./edit.php?id=$product->id'>Edit</a>
-                    <a href='./?delete&id=$product->id'>Delete</a>      
+                    <a onclick='return confirm(`Are you sure delete this data ?`)' href='./?delete&id=$product->id'>Delete</a>      
                 </td>
             </tr>
             ";
@@ -52,7 +52,7 @@ if (isset($_GET['delete'])) {
     if ($_GET['id']) {
         $result = $db->query("DELETE FROM products WHERE id = " . $_GET['id']);
         if ($result) {
-           exit( header('Location:/fullstack-php-beginner/products/'));
+           exit( header('Location:/php-beginner/products/'));
         }
     }
 }
